test: move RTL tests to e2e
